AnnotationDocument & SelectedItemWrapper: Add applyTransform()
AnnotationDocument::applyTransform() Allows applying a QMatrix4x4 to the image.
SelectedItemWrapper::applyTransform() Allows applying a QMatrix4x4 to an annotation's geometry.

We store the matrix as a QMatrix4x4 instead of a QTransform even though QTransform
is more suitable for the internal logic. There are subtle differences between
QMatrix4x4 and QTransform that can cause bugs when switching between the two,
especially when multiplying two QTransforms or QMatrix4x4s. It is better to
just stick with one and only use the other when you need to. We stick with
QMatrix4x4 because it's already usable in QML.

Also:

- Added SelectedItemWrapper::geometryPath, the real item geometry.
- Added SelectedItemWrapper::transform, the combination of all transforms applied to the item.
  - Resets when changes are committed with SelectedItemWrapper::commitChanges().
- Added functions to Utils for handling handle logic in C++ to keep the QML clean.
- Changed the drag handler logic in AnnotationSeletionTool to work with image and annotation transformations.
- Changed outlines for hover and selection to scale based on AnnotationDocument's image transform.